PP-9369 Refactor ChargeServiceTest and friends
• Convert ChargeServiceTest to JUnit 5
• Convert ChargeServiceFindTest to JUnit 5 and make it no longer
  extend ChargeServiceTest
• Convert ChargeServiceCreateTest to JUnit 5, make it no longer
  extend ChargeServiceTest and split some tests into their own
  test classes (which use JUnit 5 and do not extend another class)

No changes to the functionality of any of the tests.
